reco_data_iep/requirements.txt
fastapi
uvicorn
pydantic
mysql-connector-python
qdrant-client
google-api-python-client
google-auth
python-dotenv

